# public/api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Wavecrest Pro API",
    "version": "1.0.0",
    "description": "Wavecrest Pro is a real-time trend intelligence platform that aggregates trending topics across multiple social platforms. This API provides endpoints for authentication, trend discovery, dashboard management, billing, and analytics.",
    "contact": {
      "name": "Wavecrest Pro Support",
      "url": "https://wavecrest.pro",
      "email": "support@wavecrest.pro"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://wavecrest.pro",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/trends": {
      "get": {
        "tags": ["Trends"],
        "summary": "Get trending topics",
        "description": "Retrieve the latest trending topics across all platforms. Free users see a limited subset; paid plans unlock full results.",
        "operationId": "getTrends",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "description": "Filter trends by date (YYYY-MM-DD). Defaults to today.",
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "platform",
            "in": "query",
            "description": "Filter by platform name",
            "schema": { "type": "string", "enum": ["twitter", "reddit", "tiktok", "youtube", "google", "github"] }
          },
          {
            "name": "score",
            "in": "query",
            "description": "Minimum trend score (0-100)",
            "schema": { "type": "integer", "minimum": 0, "maximum": 100 }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of results to return (default 20, max 100)",
            "schema": { "type": "integer", "default": 20, "maximum": 100 }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Pagination offset",
            "schema": { "type": "integer", "default": 0 }
          }
        ],
        "responses": {
          "200": {
            "description": "List of trending topics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrendsResponse"
                },
                "example": {
                  "date": "2025-01-15",
                  "trends": [
                    {
                      "id": 142,
                      "topic": "Apple Vision Pro 2",
                      "score": 94,
                      "platform": "twitter",
                      "category": "technology",
                      "fetched_at": "2025-01-15T14:30:00.000Z"
                    },
                    {
                      "id": 143,
                      "topic": "Super Bowl LIX",
                      "score": 91,
                      "platform": "google",
                      "category": "sports",
                      "fetched_at": "2025-01-15T14:30:00.000Z"
                    }
                  ],
                  "total": 248,
                  "user_plan": "pro"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/trends/search": {
      "get": {
        "tags": ["Trends"],
        "summary": "Search trends by keyword",
        "description": "Full-text search across all trending topics. Requires Pro plan or higher.",
        "operationId": "searchTrends",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "description": "Search query string",
            "schema": { "type": "string" }
          },
          {
            "name": "platform",
            "in": "query",
            "description": "Filter results by platform",
            "schema": { "type": "string", "enum": ["twitter", "reddit", "tiktok", "youtube", "google", "github"] }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                },
                "example": {
                  "query": "AI",
                  "platform": null,
                  "count": 12,
                  "results": [
                    {
                      "topic": "OpenAI GPT-5 Launch",
                      "score": 97,
                      "platform": "twitter",
                      "fetched_at": "2025-01-15T14:30:00.000Z"
                    },
                    {
                      "topic": "AI Video Generation",
                      "score": 85,
                      "platform": "youtube",
                      "fetched_at": "2025-01-15T12:00:00.000Z"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Missing search query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Query parameter 'q' is required"
                }
              }
            }
          },
          "403": {
            "description": "Plan upgrade required (Pro+ only)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Search requires a Pro plan or higher"
                }
              }
            }
          }
        }
      }
    },
    "/api/trends/history": {
      "get": {
        "tags": ["Trends"],
        "summary": "Get historical trend data",
        "description": "Retrieve trend data for a specific topic over a date range. Useful for tracking how a topic's popularity changes over time.",
        "operationId": "getTrendHistory",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "topic",
            "in": "query",
            "required": true,
            "description": "Topic name to look up",
            "schema": { "type": "string" }
          },
          {
            "name": "days",
            "in": "query",
            "description": "Number of days of history (default 7, max 90)",
            "schema": { "type": "integer", "default": 7, "maximum": 90 }
          }
        ],
        "responses": {
          "200": {
            "description": "Historical data for the topic",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                },
                "example": {
                  "topic": "Apple Vision Pro 2",
                  "days": 7,
                  "history": [
                    { "date": "2025-01-09", "score": 42, "platform": "twitter" },
                    { "date": "2025-01-10", "score": 55, "platform": "twitter" },
                    { "date": "2025-01-11", "score": 68, "platform": "twitter" },
                    { "date": "2025-01-12", "score": 74, "platform": "twitter" },
                    { "date": "2025-01-13", "score": 82, "platform": "twitter" },
                    { "date": "2025-01-14", "score": 89, "platform": "twitter" },
                    { "date": "2025-01-15", "score": 94, "platform": "twitter" }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Missing topic parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Query parameter 'topic' is required"
                }
              }
            }
          }
        }
      }
    },
    "/api/trends/platforms": {
      "get": {
        "tags": ["Trends"],
        "summary": "List supported platforms",
        "description": "Returns a list of all platforms from which Wavecrest aggregates trending data, along with their current status.",
        "operationId": "getPlatforms",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "List of platforms",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformsResponse"
                },
                "example": {
                  "platforms": [
                    { "id": "twitter", "name": "Twitter / X", "status": "active", "last_fetch": "2025-01-15T14:30:00.000Z" },
                    { "id": "reddit", "name": "Reddit", "status": "active", "last_fetch": "2025-01-15T14:28:00.000Z" },
                    { "id": "tiktok", "name": "TikTok", "status": "active", "last_fetch": "2025-01-15T14:25:00.000Z" },
                    { "id": "youtube", "name": "YouTube", "status": "active", "last_fetch": "2025-01-15T14:20:00.000Z" },
                    { "id": "google", "name": "Google Trends", "status": "active", "last_fetch": "2025-01-15T14:15:00.000Z" },
                    { "id": "github", "name": "GitHub", "status": "active", "last_fetch": "2025-01-15T14:10:00.000Z" }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session cookie set after login or signup"
      }
    },
    "schemas": {
...
      "Trend": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "topic": { "type": "string", "description": "Trending topic name" },
          "score": { "type": "integer", "minimum": 0, "maximum": 100, "description": "Trend score (0-100)" },
          "platform": { "type": "string", "enum": ["twitter", "reddit", "tiktok", "youtube", "google", "github"] },
          "category": { "type": "string", "description": "Topic category" },
          "fetched_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "Platform": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "status": { "type": "string", "enum": ["active", "inactive", "maintenance"] },
          "last_fetch": { "type": "string", "format": "date-time" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "description": "Human-readable error message" }
        }
      },
      "TrendsResponse": {
        "type": "object",
        "properties": {
          "date": { "type": "string", "format": "date" },
          "trends": { "type": "array", "items": { "$ref": "#/components/schemas/Trend" } },
          "total": { "type": "integer" },
          "user_plan": { "type": "string" }
        }
      },
      "SearchResponse": {
        "type": "object",
        "properties": {
          "query": { "type": "string" },
          "platform": { "type": "string", "nullable": true },
          "count": { "type": "integer" },
          "results": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "topic": { "type": "string" },
                "score": { "type": "integer" },
                "platform": { "type": "string" },
                "fetched_at": { "type": "string", "format": "date-time" }
              }
            }
          }
        }
      },
      "HistoryResponse": {
        "type": "object",
        "properties": {
          "topic": { "type": "string" },
          "days": { "type": "integer" },
          "history": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "date": { "type": "string", "format": "date" },
                "score": { "type": "integer" },
                "platform": { "type": "string" }
              }
            }
          }
        }
      },
      "PlatformsResponse": {
        "type": "object",
        "properties": {
          "platforms": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Platform" }
          }
        }
      },
...
    }
  }
}